Bug with thread pool

+++ app/src/main/java/com/dmitryvoronko/news/model/data/Entry.java
@@ -1,10 +1,11 @@
 package com.dmitryvoronko.news.model.data;
 
 /**
+ *
  * Created by Dmitry on 03/11/2016.
  */
 
 public interface Entry extends Channel
 {
-    int getChannelId();
+    long getChannelId();
 }
